Move struct `object` into new header `kernel/object.hpp`
