+++ include/meevax/kernel/collector.hpp
@@ -17,109 +17,17 @@
 #ifndef INCLUDED_MEEVAX_KERNEL_COLLECTOR_HPP
 #define INCLUDED_MEEVAX_KERNEL_COLLECTOR_HPP
 
-#include <concepts>
-#include <memory> // std::allocator
-#include <vector>
-
-#include <meevax/concepts/any_of.hpp>
 #include <meevax/iostream/escape_sequence.hpp>
 #include <meevax/iostream/lexical_cast.hpp>
-#include <meevax/kernel/character.hpp>
-#include <meevax/kernel/instruction.hpp>
+#include <meevax/kernel/object.hpp>
 #include <meevax/memory/allocator.hpp>
 #include <meevax/memory/literal.hpp>
-#include <meevax/memory/nan_boxing_pointer.hpp>
 #include <meevax/memory/pointer_set.hpp>
+#include <memory> // std::allocator
 
 namespace meevax::inline kernel
 {
-  using null = std::nullptr_t;
-
-  using small_integer = std::int32_t; // Fixed sized integer that can be boxed.
-  using widen_integer = std::int64_t; // Fixed sized integer that is temporarily widened to prevent possible overflow.
-
   inline constexpr struct with_braces_tag {} with_braces {};
-
-    struct pair;
-
-    struct object : public nan_boxing_pointer<pair, bool, small_integer, float, character, instruction>
-    {
-      using pointer = nan_boxing_pointer<pair, bool, small_integer, float, character, instruction>;
-
-      object(std::nullptr_t = nullptr) noexcept;
-
-      object(object const&);
-
-      object(pair *);
-
-      template <any_of<bool, small_integer, float, double, character, instruction> T>
-      object(T const& datum)
-        : pointer { datum }
-      {
-        assert(pointer::get() == nullptr);
-      }
-
-      ~object();
-
-      auto operator =(object const&) -> object &;
-
-      auto operator =(std::nullptr_t) -> object &;
-
-      auto reset(object const&) -> void;
-
-      auto reset(std::nullptr_t = nullptr) -> void;
-
-      template <typename U>
-      auto static as(auto&& m) -> decltype(auto)
-      {
-        if constexpr (std::is_same_v<std::decay_t<U>, pair>)
-        {
-          return *m;
-        }
-        else if constexpr (std::is_class_v<std::decay_t<U>>)
-        {
-          if (auto data = dynamic_cast<std::add_pointer_t<U>>(m.get()); data)
-          {
-            return *data;
-          }
-          else
-          {
-            throw std::runtime_error("no viable conversion from " + demangle(m.type()) + " to " + demangle(typeid(U)));
-          }
-        }
-        else
-        {
-          return m.pointer::template as<U>();
-        }
-      }
-
-      template <typename U> auto as        ()       -> decltype(auto) { return as<U>                     (*this) ; }
-      template <typename U> auto as        () const -> decltype(auto) { return as<U>                     (*this) ; }
-      template <typename U> auto as_mutable() const -> decltype(auto) { return as<U>(const_cast<object &>(*this)); }
-
-      auto eqv(object const&) const -> bool;
-
-      template <typename U>
-      auto is() const
-      {
-        return type() == typeid(std::decay_t<U>);
-      }
-
-      template <typename U, typename = std::enable_if_t<std::is_class_v<U>>>
-      auto is_also() const
-      {
-        return dynamic_cast<std::add_pointer_t<U>>(pointer::get()) != nullptr;
-      }
-
-      auto type() const -> std::type_info const&;
-
-      auto write(std::ostream &) const -> std::ostream &;
-
-      auto friend operator <<(std::ostream & os, object const& datum) -> std::ostream &
-      {
-        return datum.write(os);
-      }
-    };
 
     struct pair : public std::pair<object, object>
     {
